+++ tests.py
@@ -8,7 +8,10 @@ from json.decoder import JSONDecodeError
 from pathlib import Path
 from unittest import TestCase
 
+from pydantic import TypeAdapter, ValidationError
+
 from courts_db import find_court, find_court_by_id
+from courts_db.models import CourtDict
 from courts_db.text_utils import strip_punc
 from courts_db.utils import db_root, load_courts_db
 
@@ -177,6 +180,41 @@ class JsonTest(CourtsDBTestCase):
             id = re.search(self.id_regex, court).group("id")
             name = re.search(self.name_regex, court).group("name")
             print(f"Issues with ({id}) -- {name}")
+
+    def test_courts_structure(self):
+        """
+        There’s no validation in the read path; Make sure that the JSON data
+        is in the shape that we expect.
+        """
+        type_adapter = TypeAdapter(list[CourtDict])
+
+        with open(
+            os.path.join(db_root, "data", "courts.json"),
+            encoding="utf-8",
+        ) as f:
+            data = f.read()
+
+        # We directly validate json data so that useful line numbers are given.
+        type_adapter.validate_json(data, extra="forbid")
+
+        courts = load_courts_db()
+        # Make test would fail when required key is missing
+        modified_court = courts[0].copy()
+        del modified_court["id"]
+        with self.assertRaises(ValidationError):
+            type_adapter.validate_python(modified_court, extra="forbid")
+
+        # Make test would fail when value is the wrong type
+        modified_court = courts[0].copy()
+        modified_court["id"] = 123
+        with self.assertRaises(ValidationError):
+            type_adapter.validate_python(modified_court, extra="forbid")
+
+        # Make test would fail if extra key was present
+        modified_court = courts[0].copy()
+        modified_court["NotAKeyThatWouldEverBePresent"] = True
+        with self.assertRaises(ValidationError):
+            type_adapter.validate_python([modified_court], extra="forbid")
 
     def test_unique_ids(self):
         """Are all court ids unique?"""